feat: diag split library
